Add function that checks existing style
Having read the instructions again, the user is supposed to be able
to access a stocked item either through its stock number (the number
in the array index) or its style number (which is, what I think a
shoe code that goes with the brand).

I made the styleNo member in each stock generate a random number in
which nothing is generated twice to kind of mimic how shoe brands
have unique codes.