Finish enterRecord function and as well as utility functions
I was able to pull through from start of lunch to dinnertime to finish
the needed functions, although a little bit sloppily. If I was going
to be the judge of such code, I think I'll be jumping here and there
to understand the gist of the whole enterRecord function.

I've also finished utility functions such as checking what stock is
currently recorded or not recorded, if there are already pairs in the
stock, or if there's already a style number in the records. I'm going
to try to refactor the code to make it recursive if I still have time
after this.