Finish overall structure of shoeStore, at least based on instruction
I think I went a little overboard with the features of this
assignment. Oh well, might as well enumerate everything that's
been added so far:

A bunch of utility functions,
A bunch of recursive functions taking and denying input,
And all the main requirements in passing the program.

Come to think of it, the entirety of the program consists of the
implementation of going back after every input and trying to read
what the user will type in. I took in a lot of unnecessary precau-
tions and, well, this is the result of my overthinking.