Fix bug with unchanging stock on hand and as well as unprinting stock
I also took the time to alter the code to make it so that I understand
the concept of global variables. I've been passing the global even
though I can fully access it without doing so. Having done so made
the program have unexpected results. Good thing it was only done
in enterRecord function.